Moved non-rule files from src/rules/ to src/
