+++ packages/package-json/src/createDirectPropertyValidityRule.ts
