+++ packages/package-json/src/directPropertyValidityRules.ts
