+++ packages/package-json/src/getPackagePropertiesOfNames.ts
