+++ packages/package-json/src/ruleTester.ts
